Issue #1404: ToolGroupApiService moved to acl package, existing ToolGroupApiServiceTest removed
